[ethan.mino] 인증 관리 Step 1, 2 (#29)
* feat: 토큰 발급 api 구현

* feat: 내 정보 조회 api 구현

* feat: 인증 관련 처리 추가

* refactor: 테스트 리팩토링

* feat: 인증된 사용자만 예약을 삭제할 수 있도록 변경

* refactor : Test 메서드, 클래스 public 키워드 제거 및 불필요 import 제거

* feat : 존재하지 않는 유저는 엑세스 토큰을 받을 수 없다. 테스트 메서드 추가

* feat : JwtTokenProvider 테스트 메서드 추가

* feat : Dto/Entity lombok @Getter, @ToString 적용

* feat : Request Dto validation 추가

* feat : GlobalExceptionHandler 추가

* refactor : 의존성 javax.validation:validation -> spring-boot-starter-validation

* fix : JwtTokenProvider.getPrincipal() 예외 처리

* feat : MemberUtil 추가

* feat : MemberTest 추가

* feat : TokenRequest validation Test 추가

* feat : Reservation 인증 관련 Test 메서드 추가

* feat : @LoginRequired 어노테이션 추가

* feat : Schedule 인증 관련 메서드 추가

* feat : Theme 인증 관련 메서드 추가

* refactor : AuthService의 checkInvalidLogin() 로직 수정

* refactor : ThemeTestUtil 추가 및 xxxxUtil -> xxxxTestUtil

* refactor : createToken() -> tokenLogin() 메서드명 수정

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,7 +1,9 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
+import nextstep.support.AuthorizationException;
 import nextstep.support.DuplicateEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
@@ -28,31 +30,35 @@ public class ReservationService {
         }
 
         List<Reservation> reservation = reservationDao.findByScheduleId(schedule.getId());
+
         if (!reservation.isEmpty()) {
             throw new DuplicateEntityException();
         }
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                reservationRequest.getUsername()
         );
 
         return reservationDao.save(newReservation);
     }
 
-    public List<Reservation> findAllByThemeIdAndDate(Long themeId, String date) {
+    public List<Reservation> findAllByThemeIdAndDate(Long themeId, String date, Member loginUser) {
         Theme theme = themeDao.findById(themeId);
         if (theme == null) {
             throw new NullPointerException();
         }
 
-        return reservationDao.findAllByThemeIdAndDate(themeId, date);
+        return reservationDao.findAllByThemeIdAndDateAndUsername(themeId, date, loginUser.getUsername());
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Member loginUser) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+        if (!reservation.getUsername().equals(loginUser.getUsername())) {
+            throw new AuthorizationException();
         }
 
         reservationDao.deleteById(id);
